// apps/sales-copilot/src/generated/models/ActivitypartiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Activitypartiesinstancetypecode = {
  0: 'NotRecurring',
  1: 'RecurringMaster',
  2: 'RecurringInstance',
  3: 'RecurringException',
  4: 'RecurringFutureException'
} as const;
export type Activitypartiesinstancetypecode = keyof typeof Activitypartiesinstancetypecode;
export const Activitypartiesparticipationtypemask = {
  1: 'Sender',
  2: 'ToRecipient',
  3: 'CCRecipient',
  4: 'BCCRecipient',
  5: 'Requiredattendee',
  6: 'Optionalattendee',
  7: 'Organizer',
  8: 'Regarding',
  9: 'Owner',
  10: 'Resource',
  11: 'Customer',
  12: 'ChatParticipant',
  13: 'Related'
} as const;
export type Activitypartiesparticipationtypemask = keyof typeof Activitypartiesparticipationtypemask;

export interface ActivitypartiesBase {
  "ActivityId@odata.bind": string;
  activitypartyid: string;
  addressused?: string;
  effort?: number;
  exchangeentryid?: string;
  externalid?: string;
  externalidtype?: string;
  participationtypemask: Activitypartiesparticipationtypemask;
  "PartyId@odata.bind"?: string;
  partyobjecttypecode: string;
  unresolvedpartyname?: string;
}

export interface Activityparties extends ActivitypartiesBase {
  addressusedemailcolumnnumber?: number;
  donotemail?: boolean;
  donotemailname: string;
  donotfax?: boolean;
  donotfaxname: string;
  donotphone?: boolean;
  donotphonename?: string;
  donotpostalmail?: boolean;
  donotpostalmailname?: string;
  instancetypecode: Activitypartiesinstancetypecode;
  instancetypecodename?: string;
  ispartydeleted: boolean;
  ispartydeletedname?: string;
  ownerid: string;
  owneridtype: string;
  owningbusinessunit?: string;
  owninguser?: string;
  participationtypemaskname?: string;
  scheduledend?: string;
  scheduledstart?: string;
  versionnumber?: number;
  activityid?: object;
  _activityid_value?: string;
  partyid?: object;
  _partyid_value?: string;
}